Create own component for annotation form normdata section
